+++ lightly/api/api_workflow_upload_metadata.py
@@ -1,5 +1,4 @@
-import concurrent
-import warnings
+
 from concurrent.futures import ThreadPoolExecutor
 from typing import Dict, List, Union
 from bisect import bisect_left
@@ -7,7 +6,7 @@ from bisect import bisect_left
 from tqdm import tqdm
 
 from lightly.api.utils import retry
-from lightly.cli._helpers import print_as_warning
+from lightly.utils.hipify import print_as_warning
 from lightly.openapi_generated.swagger_client.models.sample_update_request import SampleUpdateRequest
 from lightly.openapi_generated.swagger_client.models.configuration_entry import ConfigurationEntry
 from lightly.openapi_generated.swagger_client.models.configuration_set_request import ConfigurationSetRequest
