+++ lightly/cli/download_cli.py
@@ -15,9 +15,9 @@ from torch.utils.hipify.hipify_python import bcolors
 
 import lightly.data as data
 from lightly.cli._helpers import fix_input_path
-from lightly.cli._helpers import print_as_warning
 from lightly.cli._helpers import fix_hydra_arguments
 from lightly.cli._helpers import cpu_count
+from lightly.utils.hipify import print_as_warning
 
 from lightly.api.api_workflow_client import ApiWorkflowClient
 from lightly.openapi_generated.swagger_client import Creator
