+++ lightly/cli/lightly_cli.py
@@ -11,7 +11,7 @@ command-line interface.
 import hydra
 from omegaconf import DictConfig
 
-from lightly.cli._helpers import print_as_warning
+from lightly.utils.hipify import print_as_warning
 from lightly.cli._helpers import fix_hydra_arguments
 from lightly.cli.train_cli import _train_cli
 from lightly.cli.embed_cli import _embed_cli
